pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/technical_analysis_engine"]

[project]
name = "technical-analysis-engine"
version = "1.0.0"
description = "Self-contained Technical Analysis Engine with VectorBT integration"
authors = [
    {name = "Technical Analysis Team", email = "dev@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    # Core technical analysis engine dependencies
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "vectorbt>=0.25.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    
    # Data fetching
    "yfinance>=0.2.0",
    "requests>=2.31.0",
    
    # Visualization (compatible with vectorbt)
    "plotly>=5.17.0,<5.20.0",
]

# [project.scripts]
# Script moved to tests/examples/ - run directly from there

[project.optional-dependencies]
api = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
]

streamlit = [
    "streamlit>=1.28.0",
    "requests>=2.31.0",
]

dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "requests>=2.31.0",  # For integration tests
    "black>=23.0.0",
    "ruff>=0.1.0",
]

all = [
    "technical-analysis-engine[api,streamlit,dev]"
]

[project.urls]
"Homepage" = "https://github.com/yourusername/technical-analysis"
"Documentation" = "https://github.com/yourusername/technical-analysis/blob/main/README.md"
"Repository" = "https://github.com/yourusername/technical-analysis"
"Bug Tracker" = "https://github.com/yourusername/technical-analysis/issues"



[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
strict = true

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"